+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v10.2
 --------------------------
+- Highlight corresponded ingredient that matches a compound name or CAS number in a formula
 - Added GHS Classification for ingredient compounds
 - Added JSON export for formula making
 - For full details please refer to the CHANGELOG